+++ packages/static/src/commands/destroy.js
@@ -1,5 +1,5 @@
 const AWS = require("aws-sdk");
-const execa = require("execa");
+const ora = require("ora");
 const inquirer = require("inquirer");
 const {
   loadConfig,
@@ -8,13 +8,14 @@ const {
   getAWSWithProfile,
   monitorStack,
   deleteEnvironmentConfig,
+  emptyS3Bucket
 } = require("@designsystemsinternational/cli-utils");
 const {
   ACTION_NO_CONFIG,
-  ACTION_NO_ENV,
+  ACTION_NO_ENV
 } = require("@designsystemsinternational/cli-utils/src/constants");
 
-const destroy = async (args) => {
+const destroy = async args => {
   const { conf, packageJson } = loadConfig("static");
   const env = args && args.env ? args.env : await getEnvironment();
   const envConfig = getEnvironmentConfig(conf, env);
@@ -33,8 +34,8 @@ const destroy = async (args) => {
     {
       type: "confirm",
       name: "confirm",
-      message: `Warning! This will delete all files and resources in the ${envConfig.stack} stack. Continue?`,
-    },
+      message: `Warning! This will delete all files and resources in the ${envConfig.stack} stack. Continue?`
+    }
   ]);
 
   if (!answers.confirm) {
@@ -42,33 +43,22 @@ const destroy = async (args) => {
   }
 
   // Delete all files in the bucket
-  await execa(
-    "aws",
-    [
-      "s3",
-      "rm",
-      `s3://${envConfig.bucket}`,
-      "--profile",
-      conf.profile,
-      "--region",
-      conf.region,
-      "--recursive",
-    ],
-    {
-      stdout: "inherit",
-    }
-  );
+  const spinner = ora("Deleting files in S3 bucket").start();
+  await emptyS3Bucket(AWS, envConfig.bucket);
+  spinner.succeed();
 
   // Cloudformation!
+  spinner.start("Running deleteStack on CloudFormation");
   const cloudformation = new AWS.CloudFormation();
   await cloudformation.deleteStack({ StackName: envConfig.stack }).promise();
+  spinner.succeed();
 
   // Listen for changes to cloudformation
   await monitorStack(AWS, envConfig.stack);
 
+  spinner.start("Deleting environment config");
   deleteEnvironmentConfig("static", env);
-
-  console.log("Done!");
+  spinner.succeed();
 };
 
 module.exports = destroy;